requirements.txt
astroid==2.2.5
boto3==1.9.224
botocore==1.12.224
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.4
django-cleanup==4.0.0
django-countries==5.5
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.15.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
protobuf==3.9.1
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
stripe==2.36.2
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
